+++ app/components/mobile-dialog.hbs
@@ -1,10 +1,12 @@
 {{#if @dev}}
-  {{#if this.toShow}}
+  {{#if @deviceType}}
+    {{#if this.toShow}}
     <dialog class="mobile-dialog__dialog">
         Open RDS in app
       <br/>
       <button id="mobile-dialog__close-button" type="button" {{on 'click' (fn this.closeDialog)}} >Cancle</button>
       <button id="mobile-dialog__open-button" type="button" {{on 'click' (fn this.openRDSApp)}}>Okay</button>
     </dialog>
+    {{/if}}
   {{/if}}
 {{/if}}
