+++ app/controllers/application.js
@@ -1,5 +1,6 @@
 import Controller from '@ember/controller';
 import { inject as service } from '@ember/service';
+import { tracked } from '@glimmer/tracking';
 import { MAIN_SITE_URL } from '../constants/url';
 import { action } from '@ember/object';
 import { GITHUB_URL } from '../constants/url';
@@ -8,6 +9,7 @@ import ENV from 'website-my/config/environment';
 export default class ApplicationController extends Controller {
   @service router;
   @service featureFlag;
+  @tracked toVisible = this.checkDeviceType();
 
   GITHUB_URL = GITHUB_URL;
   BASE_API_URL = ENV.BASE_API_URL;
@@ -31,6 +33,18 @@ export default class ApplicationController extends Controller {
       }
     } catch (err) {
       console.error('Error: ', err);
+    }
+  }
+
+  checkDeviceType() {
+    let regexp = /android|iphone|kindle|ipad/i;
+    let details = navigator.userAgent;
+    let isMobileDevice = regexp.test(details);
+
+    if (isMobileDevice) {
+      return true;
+    } else {
+      return false;
     }
   }
 }
