+++ app/templates/application.hbs
@@ -2,6 +2,7 @@
 
 <MobileDialog
 @dev ={{this.isDevMode}}
+@deviceType = {{this.toVisible}}
 />
 <Navbar
   @firstName={{@model.firstName}}
